requirements.txt
torch
numpy
faiss-cpu
scipy
Levenshtein
tqdm
